Changed longs to long longs because we will be dealing with a lot of things, so I don't want an overflow

+++ PrimaryHash.cpp
@@ -152,7 +152,7 @@ void PrimaryHash::printStatistics()
 
 string PrimaryHash::find(string cityToFind)
 {
-	unsigned long key = hash(cityToFind);
+	unsigned long long key = hash(cityToFind);
 
 	cout << "city details: Key = " << key << endl;
 
@@ -214,17 +214,17 @@ void PrimaryHash::addToHash(City *city)
 	}
 }
 
-unsigned long PrimaryHash::hash(string place)
+unsigned long long PrimaryHash::hash(string place)
 {
-	unsigned long x = numerizeString(place);
+	unsigned long long x = numerizeString(place);
 
 	return (((a * x) + b) % prime2) % capacity;
 }
 
-unsigned long PrimaryHash::numerizeString(string place)
+unsigned long long PrimaryHash::numerizeString(string place)
 {
-	unsigned long sum = 0;
-	for(unsigned long i = 0; i < place.size(); i++)
+	unsigned long long sum = 0;
+	for(unsigned long long i = 0; i < place.size(); i++)
 	{
 		sum = ((sum * c) + place.at(i)) % prime1;
 	}
